+++ nav-app/src/app/data.service.ts
@@ -370,6 +370,7 @@ export class Technique extends BaseStix {
     public readonly platforms: string[];        // platforms for this technique.
     public readonly tactics: string[];          // tactics this technique is found under in phase-name format
     public readonly subtechniques: Technique[]; // subtechniques under this technique
+    public readonly datasources: string;        // data sources of the technique
     public parent: Technique = null;            // parent technique. Only present if it's a sub-technique
     public get isSubtechnique() { return this.parent != null; }
     /**
@@ -380,6 +381,10 @@ export class Technique extends BaseStix {
     constructor(stixSDO: any, subtechniques: Technique[], dataService: DataService) {
         super(stixSDO, dataService);
         this.platforms = stixSDO.x_mitre_platforms;
+      	if (stixSDO.x_mitre_data_sources !== undefined)
+		      this.datasources = stixSDO.x_mitre_data_sources.toString();
+	      else
+		      this.datasources = "";
         this.tactics = stixSDO.kill_chain_phases.map((phase) => phase.phase_name);
         this.subtechniques = subtechniques;
         for (let subtechnique of this.subtechniques) {